FlightTasks: move FlightTask to own folder/lib
fix header includes
add cmake subdirectories dynamically
